+++ lib/utopia/request.rb
@@ -4,9 +4,11 @@
 # Copyright, 2026, by Samuel Williams.
 
 require "stringio"
-
 require "protocol/http/request"
+require "protocol/url"
 require "protocol/url/form_data/parser"
+
+require_relative "path"
 
 module Utopia
 	# Utopia's application-facing request wrapper.
@@ -19,103 +21,23 @@ module Utopia
 			self.new(Protocol::HTTP::Request[*arguments])
 		end
 		
-		# Initialize the request proxy.
+		# Initialize the request.
 		# @parameter delegate [Protocol::HTTP::Request] The underlying protocol request.
-		# @parameter request_path [String | Nil] The original path before internal rewrites.
-		def initialize(delegate, request_path: nil)
+		def initialize(delegate)
 			@delegate = delegate
-			@request_path = request_path
+			
+			@url = nil
 			@session = nil
 			@variables = nil
 			@localization = nil
 			@exception = nil
 			
-			@query_arguments = nil
+			@query_parameters = nil
 			@cookies = nil
 		end
 		
 		# The underlying protocol request.
 		attr :delegate
-		
-		# Duplicate the underlying protocol request when duplicating this proxy.
-		# @parameter other [Request] The request being copied.
-		def initialize_copy(other)
-			super
-			
-			@delegate = other.delegate.dup
-			@query_arguments = nil
-			@cookies = nil
-		end
-		
-		# Assign the request path including query string.
-		def path= value
-			if value != @delegate.path
-				@request_path ||= self.path_info
-			end
-			
-			@delegate.path = value
-			@query_arguments = nil
-		end
-		
-		# Whether the request method is POST.
-		def post?
-			self.method == "POST"
-		end
-		
-		# The request path without the query string.
-		def path_info
-			self.path&.split("?", 2)&.first
-		end
-		
-		# Set the request path while preserving the query string.
-		def path_info= value
-			@request_path ||= self.path_info
-			
-			if query = self.query
-				self.path = "#{value}?#{query}"
-			else
-				self.path = value
-			end
-		end
-		
-		# The original request path, before any internal request rewrites.
-		def request_path
-			@request_path || self.path_info
-		end
-		
-		# The query string without the leading question mark.
-		def query
-			path = self.path
-			
-			if path&.include?("?")
-				return path.split("?", 2).last
-			end
-		end
-		
-		# Decoded query arguments.
-		def query_arguments
-			@query_arguments ||= decode_arguments(self.query)
-		end
-		
-		# Decoded request cookies.
-		def cookies
-			@cookies ||= parse_cookies(self.headers["cookie"])
-		end
-		
-		# The request host with optional port.
-		def host
-			self.authority || self.headers["host"]
-		end
-		
-		# The request user agent.
-		def user_agent
-			self.headers["user-agent"]
-		end
-		
-		# The request referrer.
-		def referrer
-			self.headers["referer"]
-		end
 		
 		# The session associated with this request, if installed.
 		attr_accessor :session
@@ -129,40 +51,135 @@ module Utopia
 		# The exception associated with this request, if any.
 		attr_accessor :exception
 		
-		# The remote peer IP address, if available.
-		def ip
-			self.peer&.ip_address
-		end
-		
-		# The full request URL, if scheme and host are available.
-		def url
-			if scheme = self.scheme and host = self.host
-				"#{scheme}://#{host}#{self.path}"
+		private def parse_url
+			path, query = @delegate.path.split("?", 2)
+			
+			if scheme = @delegate.scheme and authority = @delegate.authority
+				Protocol::URL::Absolute.new(scheme, authority, path, query)
 			else
-				self.path
+				Protocol::URL::Relative.new(path, query)
 			end
 		end
 		
-		# Build a derived request with updated protocol fields.
-		def with(method: self.method, path: self.path, path_info: nil)
-			delegate = @delegate.dup
-			delegate.method = method
+		# The structured request URL.
+		# @returns [Protocol::URL::Absolute | Protocol::URL::Relative | Nil] The request URL.
+		def url
+			@url ||= self.parse_url.normalize!
+		end
+		
+		# @returns [Utopia::Path] The path components.
+		def path
+			Path.new(self.url.path.components(Protocol::URL::Encoding::System))
+		end
+		
+		# Assign the decoded application path while preserving the current query string.
+		# @parameter value [Utopia::Path | Protocol::URL::Path | String] The path.
+		def path= value
+			if value.is_a?(Protocol::URL::Path)
+				self.url.path = value
+			else
+				self.url.path = Path[value].to_url_path
+			end
+		end
+		
+		# Whether the request method is QUERY.
+		# @returns [Boolean] Whether the request method is QUERY.
+		def query?
+			@delegate.method == "QUERY"
+		end
+		
+		# Whether the request method is POST.
+		def post?
+			@delegate.method == "POST"
+		end
+		
+		# The original request path, before any internal request rewrites.
+		def request_path
+			@delegate.path.split("?", 2).first
+		end
+		
+		# The original request path as decoded application components.
+		# @returns [Utopia::Path] The normalized original request path.
+		def original_path
+			path = Protocol::URL::Path[self.request_path].normalize
+			return Path.new(path.components(Protocol::URL::Encoding::System))
+		end
+		
+		QUERY_PARSER = Protocol::URL::FormData::Parser.new
+		
+		private def parse_query_parameters(query)
+			if query
+				return QUERY_PARSER.parse(StringIO.new(query))
+			else
+				Hash.new
+			end
+		end
+		
+		# Decoded query arguments.
+		def query_parameters
+			@query_parameters ||= parse_query_parameters(self.url.query)
+		end
+		
+		private def parse_cookies(cookie_header)
+			cookies = {}
 			
-			request = self.class.new(delegate, request_path: self.request_path)
+			return cookies unless cookie_header
+			
+			if cookie_header.respond_to?(:to_str)
+				cookie_header = cookie_header.to_str
+			else
+				cookie_header = cookie_header.to_s
+			end
+			
+			cookie_header.split(/;\s*/).each do |pair|
+				key, value = pair.split("=", 2)
+				cookies[key] = value || ""
+			end
+			
+			return cookies
+		end
+		
+		# Decoded request cookies.
+		def cookies
+			@cookies ||= parse_cookies(self.headers["cookie"])
+		end
+		
+		# The request user agent.
+		def user_agent
+			@delegate.headers["user-agent"]
+		end
+		
+		# The request referrer.
+		def referrer
+			@delegate.headers["referer"]
+		end
+		
+		# The remote peer IP address, if available.
+		def ip
+			@delegate.peer&.ip_address
+		end
+		
+		# Assign the structured request URL.
+		# @parameter url [Protocol::URL::Absolute | Protocol::URL::Relative | String | Nil] The request URL.
+		def url= url
+			@url = Protocol::URL[url]
+		end
+		
+		# Build a derived request with updated protocol fields.
+		def with(method: nil, path: self.path)
+			delegate = @delegate
+			
+			if method
+				delegate = @delegate.dup
+				delegate.method = method
+			end
+			
+			request = self.class.new(delegate)
 			request.session = @session
 			request.variables = @variables
 			request.localization = @localization
 			request.exception = @exception
-			
-			if path_info
-				if query = self.query
-					request.path = "#{path_info}?#{query}"
-				else
-					request.path = path_info
-				end
-			else
-				request.path = path
-			end
+			request.path = path
 			
 			return request
 		end
@@ -182,32 +199,6 @@ module Utopia
 		
 		def respond_to_missing?(name, include_private = false)
 			@delegate.respond_to?(name) || super(name, include_private)
-		end
-		
-		def decode_arguments(query)
-			return {} unless query
-			
-			parser = Protocol::URL::FormData::Parser.new
-			return parser.parse(StringIO.new(query))
-		end
-		
-		def parse_cookies(cookie_header)
-			cookies = {}
-			
-			return cookies unless cookie_header
-			
-			if cookie_header.respond_to?(:to_str)
-				cookie_header = cookie_header.to_str
-			else
-				cookie_header = cookie_header.to_s
-			end
-			
-			cookie_header.split(/;\s*/).each do |pair|
-				key, value = pair.split("=", 2)
-				cookies[key] = value || ""
-			end
-			
-			return cookies
 		end
 	end
 end